Refactor to separate hooks
